+++ src/main/java/com/enjin/enjincoin/sdk/service/ethereum/AsynchronousEthereumService.java
@@ -2,8 +2,10 @@ package com.enjin.enjincoin.sdk.service.ethereum;
 
 import com.enjin.enjincoin.sdk.http.Callback;
 
+import java.math.BigInteger;
+
 public interface AsynchronousEthereumService {
 
-    void getAllowanceAsync(String ethAddr, Callback<Integer> callback);
+    void getAllowanceAsync(String ethAddr, Callback<BigInteger> callback);
 
 }
