+++ src/main/java/com/enjin/enjincoin/sdk/service/ethereum/SynchronousEthereumService.java
@@ -3,9 +3,10 @@ package com.enjin.enjincoin.sdk.service.ethereum;
 import com.enjin.enjincoin.sdk.http.Result;
 
 import java.io.IOException;
+import java.math.BigInteger;
 
 public interface SynchronousEthereumService {
 
-    Result<Integer> getAllowanceSync(String ethAddr) throws IOException;
+    Result<BigInteger> getAllowanceSync(String ethAddr) throws IOException;
 
 }
