+++ src/main/java/com/enjin/enjincoin/sdk/service/ethereum/impl/EthereumRetrofitService.java
@@ -4,9 +4,11 @@ import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.Path;
 
+import java.math.BigInteger;
+
 public interface EthereumRetrofitService {
 
     @GET("api/v1/ethereum/{eth_addr}/allowance")
-    Call<Integer> getAllowance(@Path("eth_addr") String ethAddr);
+    Call<BigInteger> getAllowance(@Path("eth_addr") String ethAddr);
 
 }
